refactor: extract shared ui classes
